Separate out archiver.rs
